requirements.txt
python-telegram-bot[job-queue,web,webhooks]>=20.0
instaloader>=4.10
python-dotenv>=1.0.0
aiohttp>=3.8.0
